+++ app/components/preprint-status-banner/component.js
@@ -8,6 +8,8 @@ import latestAction from 'reviews/utils/latest-action';
 const PENDING = 'pending';
 const ACCEPTED = 'accepted';
 const REJECTED = 'rejected';
+const PENDING_WITHDRAWAL = 'pending_withdrawal';
+const WITHDRAWN = 'withdrawn';
 
 const PRE_MODERATION = 'pre-moderation';
 const POST_MODERATION = 'post-moderation';
@@ -15,15 +17,19 @@ const POST_MODERATION = 'post-moderation';
 const COMMENT_LIMIT = 65535;
 
 const ICONS = {
-    [PENDING]: 'fa-hourglass-o',
-    [ACCEPTED]: 'fa-check-circle-o',
-    [REJECTED]: 'fa-times-circle-o',
+    [PENDING]: ['fa-hourglass-o'],
+    [ACCEPTED]: ['fa-check-circle-o'],
+    [REJECTED]: ['fa-times-circle-o'],
+    [PENDING_WITHDRAWAL]: ['fa-hourglass-o'],
+    [WITHDRAWN]: ['fa-circle-o', 'fa-minus inner-icon'],
 };
 
 const STATUS = {
     [PENDING]: 'components.preprintStatusBanner.pending',
     [ACCEPTED]: 'components.preprintStatusBanner.accepted',
     [REJECTED]: 'components.preprintStatusBanner.rejected',
+    [PENDING_WITHDRAWAL]: 'components.preprintStatusBanner.pending',
+    [WITHDRAWN]: 'components.preprintStatusBanner.withdrawn',
 };
 
 const MESSAGE = {
@@ -31,6 +37,8 @@ const MESSAGE = {
     [POST_MODERATION]: 'components.preprintStatusBanner.message.pendingPost',
     [ACCEPTED]: 'components.preprintStatusBanner.message.accepted',
     [REJECTED]: 'components.preprintStatusBanner.message.rejected',
+    [PENDING_WITHDRAWAL]: 'components.preprintStatusBanner.message.pendingWithdrawal',
+    [WITHDRAWN]: 'components.preprintStatusBanner.message.withdrawn',
 };
 
 const CLASS_NAMES = {
@@ -38,6 +46,8 @@ const CLASS_NAMES = {
     [POST_MODERATION]: 'preprint-status-pending-post',
     [ACCEPTED]: 'preprint-status-accepted',
     [REJECTED]: 'preprint-status-rejected',
+    [PENDING_WITHDRAWAL]: 'preprint-status-pending-withdrawal',
+    [WITHDRAWN]: 'preprint-status-withdrawn',
 };
 
 const SETTINGS = {
@@ -79,12 +89,17 @@ const DECISION_EXPLANATION = {
         [PRE_MODERATION]: 'components.preprintStatusBanner.decision.reject.pre',
         [POST_MODERATION]: 'components.preprintStatusBanner.decision.reject.post',
     },
+    withdrawn: {
+        [POST_MODERATION]: 'components.preprintStatusBanner.decision.withdrawn.post',
+    },
 };
 
 const RECENT_ACTIVITY = {
     [PENDING]: 'components.preprintStatusBanner.recentActivity.pending',
     [ACCEPTED]: 'components.preprintStatusBanner.recentActivity.accepted',
     [REJECTED]: 'components.preprintStatusBanner.recentActivity.rejected',
+    [PENDING_WITHDRAWAL]: 'components.preprintStatusBanner.recentActivity.pendingWithdrawal',
+    [WITHDRAWN]: 'components.preprintStatusBanner.recentActivity.withdrawn',
     automatic: {
         [PENDING]: 'components.preprintStatusBanner.recentActivity.automatic.pending',
         [ACCEPTED]: 'components.preprintStatusBanner.recentActivity.automatic.accepted',
@@ -101,6 +116,10 @@ export default Component.extend({
     commentPlaceholder: 'components.preprintStatusBanner.decision.commentPlaceholder',
     labelAccept: 'components.preprintStatusBanner.decision.accept.label',
     labelReject: 'components.preprintStatusBanner.decision.reject.label',
+    labelApprove: 'components.preprintStatusBanner.decision.approve.label',
+    labelDecline: 'components.preprintStatusBanner.decision.decline.label',
+    approveExplanation: 'components.preprintStatusBanner.decision.approve.explanation',
+    declineExplanation: 'components.preprintStatusBanner.decision.decline.explanation',
 
     classNames: ['preprint-status-component'],
 
@@ -112,6 +131,7 @@ export default Component.extend({
     reviewerComment: '',
     decision: ACCEPTED,
     decisionValueToggled: false,
+    withdrawalJustification: alias('withdrawalRequest.comment'),
 
     reviewsWorkflow: alias('submission.provider.reviewsWorkflow'),
     reviewsCommentsPrivate: alias('submission.provider.reviewsCommentsPrivate'),
@@ -119,10 +139,36 @@ export default Component.extend({
 
     creatorProfile: alias('latestAction.creator.profileURL'),
     creatorName: alias('latestAction.creator.fullName'),
+    withdrawalRequesterProfile: alias('withdrawalRequest.creator.profileURL'),
+    withdrawalRequesterName: alias('withdrawalRequest.creator.fullName'),
 
     commentExceedsLimit: computed.gt('reviewerComment.length', COMMENT_LIMIT),
 
     userActivity: computed.or('commentEdited', 'decisionValueToggled'),
+
+    labelDate: computed('submission', function() {
+        return this.get('submission.dateWithdrawn') ?
+            this.get('submission.dateWithdrawn') :
+            this.get('submission.dateLastTransitioned');
+    }),
+
+    labelDecision: computed('submission', function() {
+        return this.get('submission.isPublished') ?
+            'components.preprintStatusBanner.decision.withdrawn.label' :
+            'components.preprintStatusBanner.decision.reject.label';
+    }),
+
+    radioDecision: computed('submission', function() {
+        return this.get('submission.isPublished') ?
+            'withdrawn' :
+            'rejected';
+    }),
+
+    isDisabled: computed('latestAction', 'submission.reviewActions.isPending', function() {
+        const reason = this.get('latestAction.comment');
+        const type = this.get('submission.reviewsState');
+        return ((type === 'withdrawn' && !reason) || this.get('submission.reviewActions.isPending'));
+    }),
 
     commentLengthErrorMessage: computed('reviewerComment', function () {
         const i18n = this.get('i18n');
@@ -133,21 +179,37 @@ export default Component.extend({
     }),
 
 
-    statusExplanation: computed('reviewsWorkflow', 'submission.reviewsState', function() {
+    statusExplanation: computed('reviewsWorkflow', 'submission.reviewsState', 'isPendingWithdrawal', function() {
+        if (this.get('isPendingWithdrawal')) {
+            return MESSAGE[PENDING_WITHDRAWAL];
+        }
         return this.get('submission.reviewsState') === PENDING ?
             MESSAGE[this.get('reviewsWorkflow')] :
             MESSAGE[this.get('submission.reviewsState')];
     }),
 
-    status: computed('submission.reviewsState', function() {
-        return STATUS[this.get('submission.reviewsState')];
+    status: computed('submission.reviewsState', 'isPendingWithdrawal', function() {
+        if (this.get('isPendingWithdrawal')) {
+            return STATUS[PENDING_WITHDRAWAL];
+        } else {
+            return STATUS[this.get('submission.reviewsState')];
+        }
     }),
 
-    icon: computed('submission.reviewsState', function() {
+    icon: computed('submission.reviewsState', 'isPendingWithdrawal', function() {
+        if (this.get('isPendingWithdrawal')) {
+            return ICONS[PENDING_WITHDRAWAL];
+        }
         return ICONS[this.get('submission.reviewsState')];
     }),
 
-    recentActivityLanguage: computed('noActions', 'submission.reviewsState', function() {
+    requestActivityLanguage: computed('isPendingWithdrawal', function() {
+        if (this.get('isPendingWithdrawal')) {
+            return RECENT_ACTIVITY[PENDING_WITHDRAWAL];
+        }
+    }),
+
+    recentActivityLanguage: computed('noActions', 'submission.reviewsState', 'isPendingWithdrawal', function() {
         if (this.get('noActions')) {
             return RECENT_ACTIVITY.automatic[this.get('submission.reviewsState')];
         } else {
@@ -156,14 +218,21 @@ export default Component.extend({
     }),
 
 
-    getClassName: computed('reviewsWorkflow', 'submission.reviewsState', function() {
-        return this.get('submission.reviewsState') === PENDING ?
-            CLASS_NAMES[this.get('reviewsWorkflow')] :
-            CLASS_NAMES[this.get('submission.reviewsState')];
+    getClassName: computed('reviewsWorkflow', 'submission.reviewsState', 'isPendingWithdrawal', function() {
+        if (this.get('isPendingWithdrawal')) {
+            return CLASS_NAMES[PENDING_WITHDRAWAL];
+        } else {
+            return this.get('submission.reviewsState') === PENDING ?
+                CLASS_NAMES[this.get('reviewsWorkflow')] :
+                CLASS_NAMES[this.get('submission.reviewsState')];
+        }
     }),
 
     latestAction: computed('submission.reviewActions.[]', function() {
-        return latestAction(this.get('submission.reviewActions'));
+        if (this.get('submission.reviewActions.[]')) {
+            return latestAction(this.get('submission.reviewActions'));
+        }
+        return null;
     }),
 
     noComment: computed('reviewerComment', function() {
@@ -197,22 +266,45 @@ export default Component.extend({
     acceptExplanation: computed('reviewsWorkflow', function() {
         return DECISION_EXPLANATION.accept[this.get('reviewsWorkflow')];
     }),
+
     rejectExplanation: computed('reviewsWorkflow', function() {
-        return DECISION_EXPLANATION.reject[this.get('reviewsWorkflow')];
+        if (this.get('reviewsWorkflow') === 'pre-moderation') {
+            if (this.get('submission.reviewsState') === ACCEPTED) {
+                return this.get('approveExplanation');
+            } else {
+                return DECISION_EXPLANATION.reject[this.get('reviewsWorkflow')];
+            }
+        } else {
+            return DECISION_EXPLANATION.withdrawn[this.get('reviewsWorkflow')];
+        }
     }),
 
-    labelDecisionDropdown: computed('submission.reviewsState', function() {
-        return this.get('submission.reviewsState') === PENDING ?
-            'components.preprintStatusBanner.decision.makeDecision' :
-            'components.preprintStatusBanner.decision.modifyDecision';
+    labelDecisionDropdown: computed('submission.reviewsState', 'isPendingWithdrawal', function() {
+        if (this.get('submission.reviewsState') === 'withdrawn') {
+            return 'components.preprintStatusBanner.decision.withdrawalReason';
+        } else if (this.get('isPendingWithdrawal')) {
+            return 'components.preprintStatusBanner.decision.makeDecision';
+        } else {
+            return this.get('submission.reviewsState') === PENDING ?
+                'components.preprintStatusBanner.decision.makeDecision' :
+                'components.preprintStatusBanner.decision.modifyDecision';
+        }
     }),
-    labelDecisionHeader: computed('submission.reviewsState', function() {
-        return this.get('submission.reviewsState') === PENDING ?
-            'components.preprintStatusBanner.decision.header.submitDecision' :
-            'components.preprintStatusBanner.decision.header.modifyDecision';
+    labelDecisionHeader: computed('submission.reviewsState', 'isPendingWithdrawal', function() {
+        if (this.get('submission.reviewsState') === 'withdrawn') {
+            return 'components.preprintStatusBanner.decision.header.withdrawalReason';
+        } else if (this.get('isPendingWithdrawal')) {
+            return 'components.preprintStatusBanner.decision.header.submitDecision';
+        } else {
+            return this.get('submission.reviewsState') === PENDING ?
+                'components.preprintStatusBanner.decision.header.submitDecision' :
+                'components.preprintStatusBanner.decision.header.modifyDecision';
+        }
     }),
-    labelDecisionBtn: computed('submission.reviewsState', 'decision', 'reviewerComment', function() {
-        if (this.get('submission.reviewsState') === PENDING) {
+    labelDecisionBtn: computed('submission.reviewsState', 'decision', 'reviewerComment', 'isPendingWithdrawal', function() {
+        if (this.get('isPendingWithdrawal')) {
+            return 'components.preprintStatusBanner.decision.btn.submitDecision';
+        } else if (this.get('submission.reviewsState') === PENDING) {
             return 'components.preprintStatusBanner.decision.btn.submitDecision';
         } else if (this.get('submission.reviewsState') !== this.get('decision')) {
             return 'components.preprintStatusBanner.decision.btn.modifyDecision';
@@ -245,13 +337,47 @@ export default Component.extend({
         submit() {
             let trigger = '';
             if (this.get('submission.reviewsState') !== PENDING && (this.get('commentEdited') && !this.get('decisionChanged'))) {
+                // If the submission is not pending,
+                // the decision has not changed and the comment is edited.
+                // the trigger would be 'edit_comment'
                 trigger = 'edit_comment';
             } else {
-                trigger = this.get('decision') === ACCEPTED ? 'accept' : 'reject';
+                // Otherwise, meaning that if the moderator is not editing the comment
+                let actionType = '';
+                if (this.get('submission.isPublished') && this.get('isPendingWithdrawal')) {
+                    // if the submission is published and is pending withdrawal.
+                    // actionType would be 'reject'
+                    // meaning moderators could accept/reject the withdrawl request
+                    actionType = 'reject';
+                } else if (this.get('submission.isPublished') && !this.get('isPendingWithdrawal')) {
+                    // if the submission is published and is not pending withdrawal
+                    // actionType would be 'withdraw'
+                    // meaning moderators could approve/directly withdraw the submission
+                    actionType = 'withdraw';
+                } else {
+                    // Otherwise
+                    // actionType would be 'reject'
+                    // meaning the moderator could either accept or reject the submission
+                    actionType = 'reject';
+                }
+                // If the decision is to accept the submission or the withdrawal request,
+                // the trigger is 'accept'
+                // If not, then the trigger is whatever 'actionType' set above.
+                trigger = this.get('decision') === ACCEPTED ? 'accept' : actionType;
             }
 
-            const comment = this.get('reviewerComment').trim();
-            this.get('submitDecision')(trigger, comment, 'pending');
+            let comment = '';
+            if (trigger === 'accept' && this.get('isPendingWithdrawal')) {
+                comment = this.get('withdrawalJustification').trim();
+            } else {
+                comment = this.get('reviewerComment').trim();
+            }
+
+            if (this.get('isPendingWithdrawal')) {
+                this.get('submitRequestsDecision')(trigger, comment, 'pending');
+            } else {
+                this.get('submitReviewsDecision')(trigger, comment, 'pending');
+            }
         },
         cancel() {
             this.set('decision', this.get('submission.reviewsState'));